fix(telemetry): address #1560 review — token-attribution correctness + schema hardening
Resolves CodeRabbit + Sentry/Seer findings on the #1525 W2 token-telemetry code,
evaluated against the .exarchos/invariants.md catalog:

- resolveTeammateByWorktree now selects the most-recent (timestamp, sequence)
  worktree match across all streams instead of the first — prevents
  misattributing tokens to a stale stream when a worktree path is reused across
  features over time (INV-1 correct stream attribution).
- SubagentStop payload parsed via a Zod safeParse schema; fail-open preserved
  (malformed input → { continue: true }, no throw, no block) (INV-5a/5b).
- subagent.tokens_used.outputTokens constrained to .int().nonnegative() — token
  counts are integers; fractional values would corrupt downstream aggregates.
- code-quality mutation-score trend folds ONLY the mutation-adequacy gate's
  score (gated on gateName, not mere presence of a numeric field) — prevents
  cross-gate contamination (INV-1 left-fold correctness).
- team-performance avgOutputTokensPerTask renamed → avgOutputTokensPerRun; the
  denominator is subagent runs, so the name now matches behavior (INV-5b).
- SubagentStop skips emitting a zero-token atom: it carries no usage signal and
  the run is already recorded by team.teammate.dispatched / team.task.completed,
  so suppression keeps token metrics clean without losing provenance (INV-1).

MCP suite 7756 green (+4 new tests: latest-wins resolution, zero-token skip,
malformed-payload fail-open, non-mutation-gate score ignored, fractional-token
rejection); typecheck clean.

+++ servers/exarchos-mcp/src/event-store/schemas.ts
@@ -1005,7 +1005,7 @@ export type TurnCompletedData = z.infer<typeof TurnCompletedDataSchema>;
 // `.passthrough()` keeps room for future per-subagent samples (cache tokens, etc).
 export const SubagentTokensUsedDataSchema = z.object({
   agentId: z.string().min(1).describe('Stable subagent invocation id from the SubagentStop hook (agent_id).'),
-  outputTokens: z.number().nonnegative().describe('Summed output tokens across the subagent\'s own transcript.'),
+  outputTokens: z.number().int().nonnegative().describe('Summed output tokens across the subagent\'s own transcript.'),
   agentType: z.string().optional().describe('Subagent type/name (agent_type), e.g. "exarchos-implementer".'),
   teammateName: z.string().optional().describe('Resolved teammate name when the subagent cwd matched a dispatched worktree.'),
   taskId: z.string().optional().describe('Resolved task id from the matching team dispatch.'),
